old changes

+++ src/Engine.cpp
@@ -9,7 +9,7 @@ extern "C" {
     __attribute__ ((dllexport)) unsigned long NvOptimusEnablement = 1;
     __attribute__ ((dllexport)) int AmdPowerXpressRequestHighPerformance = 1;
 #else
-	_declspec(dllexport) DWORD NvOptimusEnablement = 1;
+	_declspec(dllexport) unsigned long  NvOptimusEnablement = 1;
 	_declspec(dllexport) int AmdPowerXpressRequestHighPerformance = 1;
 #endif
 }
